pyproject.toml
[project]
name = "mini-jira-clone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cors-headers>=4.7.0",
    "drf-spectacular>=0.28.0",
    "pillow>=11.2.1",
]
